requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
PyJWT
cloudinary
flask-cors
python-dotenv